Moving decoupled store code to storeutils for clarity
The Listing and Get/Reading code depends on some cloudstorage primitives
but unrelated functionality wise since it doesn't use the Store
interface implementations.
