+++ store.go
@@ -1,9 +1,7 @@
 package cloudstorage
 
 import (
-	"bytes"
 	"fmt"
-	"io"
 	"mime"
 	"net/http"
 	"os"
@@ -80,132 +78,6 @@ func gcsCommonClient(client *http.Client, csctx *CloudStoreContext) (Store, erro
 		return nil, err
 	}
 	return store, nil
-}
-
-func GetObject(gc *storage.Client, bucket, name string) (*bytes.Buffer, error) {
-	// Get buckethandler
-	gsbh := gc.Bucket(bucket)
-
-	// Create Query
-	q := storage.Query{Prefix: name, MaxResults: 1}
-
-	// Get list of *the* object
-	ol, err := List(gsbh, q)
-	if err != nil {
-		return nil, err
-	}
-
-	buff, errs := OpenObject(ol, gsbh)
-	if buff == nil {
-		if len(errs) >= 0 {
-			i := 0
-			errBuff := bytes.NewBufferString("GetObject Errors:\n")
-			for _, e := range errs {
-				i++
-				errBuff.WriteString(e.Error())
-				errBuff.WriteString("\n")
-			}
-			err := fmt.Errorf("%s", errBuff.String())
-			return nil, err
-		} else {
-			return nil, fmt.Errorf("GetObject recieved nil byte buffer and no errors")
-		}
-	} else {
-		//success
-		return buff, nil
-	}
-}
-
-// Opens the first object returned in a storage.ObjectList
-// returns contents via byte Buffer
-func OpenObject(objects *storage.ObjectList, gcsb *storage.BucketHandle) (*bytes.Buffer, []error) {
-	var buff *bytes.Buffer = bytes.NewBuffer([]byte{})
-	log := logging.NewStdLogger(true, 4, "OpenObject")
-	var googleObject *storage.ObjectAttrs
-	errs := make([]error, 0)
-
-	for try := 0; try < GCSRetries; try++ {
-		if objects.Results != nil && len(objects.Results) != 0 {
-			googleObject = objects.Results[0]
-		}
-
-		if googleObject != nil {
-			//we have a preexisting object, so lets download it..
-			rc, err := gcsb.Object(googleObject.Name).NewReader(context.Background())
-			if err != nil {
-				errs = append(errs, fmt.Errorf("error storage.NewReader err=%v", err))
-				log.Debugf("%v", errs)
-				backoff(try)
-				continue
-			}
-			defer rc.Close()
-
-			_, err = io.Copy(buff, rc)
-			if err != nil {
-				errs = append(errs, fmt.Errorf("error coping bytes. err=%v", err))
-				log.Debugf("%v", errs)
-				backoff(try)
-				continue
-			}
-		}
-		return buff, nil
-	}
-	errs = append(errs, fmt.Errorf("OpenObject errors past limit!"))
-	return nil, errs
-}
-
-// Iterates through pages of results from a Google Storage Bucket and
-// lists objects which match the specified storage.Query
-func List(gcsb *storage.BucketHandle, query storage.Query) (*storage.ObjectList, error) {
-
-	gobjects, err := ListBucketObjectsReq(gcsb, &query, GCSRetries)
-	if err != nil {
-		fmt.Errorf("couldn't list objects. prefix=%s err=%v", query.Prefix, err)
-		return nil, err
-	}
-
-	if gobjects == nil {
-		return nil, nil
-	}
-
-	if gobjects.Next != nil {
-		q := gobjects.Next
-		for q != nil {
-			gobjectsB, err := ListBucketObjectsReq(gcsb, q, GCSRetries)
-			if err != nil {
-				fmt.Errorf("couldn't list the remaining pages of objects. prefix=%s err=%v", q.Prefix, err)
-				return nil, err
-			}
-
-			concatGCSObjects(gobjects, gobjectsB)
-
-			if gobjectsB != nil {
-				q = gobjectsB.Next
-			} else {
-				q = nil
-			}
-		}
-	}
-
-	return gobjects, nil
-}
-
-//ListObjects is a wrapper around storeage.ListObjects, that retries on a GCS error.  GCS isn't a prefect system :p, and returns an error
-//  about once every 2 weeks.
-func ListBucketObjectsReq(gcsb *storage.BucketHandle, q *storage.Query, retries int) (*storage.ObjectList, error) {
-	var lasterr error = nil
-	//GCS sometimes returns a 500 error, so we'll just retry...
-	for i := 0; i < retries; i++ {
-		objects, err := gcsb.List(context.Background(), q)
-		if err != nil {
-			fmt.Errorf("error listing objects for the bucket. try:%d q.prefix:%v err:%v", i, q.Prefix, err)
-			lasterr = err
-			backoff(i)
-			continue
-		}
-		return objects, nil
-	}
-	return nil, lasterr
 }
 
 const ContextTypeKey = "content_type"
